+++ packages/libs/accessibility/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @lssm/lib.accessibility
 
+## 1.42.2
+
+### Patch Changes
+
+- 1f9ac4c: fix
+- Updated dependencies [1f9ac4c]
+  - @lssm/lib.design-system@1.42.2
+  - @lssm/lib.ui-kit@1.42.2
+  - @lssm/lib.ui-kit-web@1.42.2
+
+## 1.42.1
+
+### Patch Changes
+
+- f043995: Fix release
+- Updated dependencies [f043995]
+  - @lssm/lib.design-system@1.42.1
+  - @lssm/lib.ui-kit-web@1.42.1
+  - @lssm/lib.ui-kit@1.42.1
+
+## 1.42.0
+
+### Minor Changes
+
+- 8eefd9c: initial release
+
+### Patch Changes
+
+- Updated dependencies [8eefd9c]
+  - @lssm/lib.design-system@1.42.0
+  - @lssm/lib.ui-kit@1.42.0
+  - @lssm/lib.ui-kit-web@1.42.0
+
 ## 1.12.0
 
 ### Minor Changes