+++ packages/libs/cost-tracking/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lssm/lib.cost-tracking
 
+## 1.42.2
+
+### Patch Changes
+
+- 1f9ac4c: fix
+
+## 1.42.1
+
+### Patch Changes
+
+- f043995: Fix release
+
+## 1.42.0
+
+### Minor Changes
+
+- 8eefd9c: initial release
+
 ## 4.0.0
 
 ### Minor Changes