+++ packages/libs/testing/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @lssm/lib.testing
 
+## 1.42.2
+
+### Patch Changes
+
+- 1f9ac4c: fix
+- Updated dependencies [1f9ac4c]
+  - @lssm/lib.contracts@1.42.2
+  - @lssm/lib.schema@1.42.2
+
+## 1.42.1
+
+### Patch Changes
+
+- f043995: Fix release
+- Updated dependencies [f043995]
+  - @lssm/lib.contracts@1.42.1
+  - @lssm/lib.schema@1.42.1
+
+## 1.42.0
+
+### Minor Changes
+
+- 8eefd9c: initial release
+
+### Patch Changes
+
+- Updated dependencies [8eefd9c]
+  - @lssm/lib.contracts@1.42.0
+  - @lssm/lib.schema@1.42.0
+
 ## 0.5.0
 
 ### Minor Changes