+++ packages/modules/lifecycle-core/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @lssm/module.lifecycle-core
 
+## 1.42.2
+
+### Patch Changes
+
+- 1f9ac4c: fix
+- Updated dependencies [1f9ac4c]
+  - @lssm/lib.lifecycle@1.42.2
+
+## 1.42.1
+
+### Patch Changes
+
+- f043995: Fix release
+- Updated dependencies [f043995]
+  - @lssm/lib.lifecycle@1.42.1
+
+## 1.42.0
+
+### Minor Changes
+
+- 8eefd9c: initial release
+
+### Patch Changes
+
+- Updated dependencies [8eefd9c]
+  - @lssm/lib.lifecycle@1.42.0
+
 ## 0.4.0
 
 ### Minor Changes