+++ packages/tools/tsdown/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lssm/tool.tsdown
 
+## 1.42.2
+
+### Patch Changes
+
+- 1f9ac4c: fix
+
+## 1.42.1
+
+### Patch Changes
+
+- f043995: Fix release
+
+## 1.42.0
+
+### Minor Changes
+
+- 8eefd9c: initial release
+
 ## 0.13.0
 
 ### Minor Changes